Move main() to cli.py
* Adapt test cases
* Adapt setup.py to create entry points from rstxml2db.cli
* Adapt .bumpversion.cfg
* Remove version.py
* Adapt __init__.py:
  * Add more variables in __init__.py like __name__,
    __url__, __license__, and __summary__
  * Move main() to cli.py
  * Move __author__ and __version__
