+++ .bumpversion.cfg
@@ -5,7 +5,7 @@ tag = False
 
 [bumpversion:file:setup.py]
 
-[bumpversion:file:src/rstxml2db/version.py]
+[bumpversion:file:src/rstxml2db/__init__.py]
 
 [bumpversion:file:doc/index.rst]
