+++ tests/test_filenotfound.py
@@ -15,7 +15,7 @@ def test_filenotfound1(args):
 
 def test_filenotfound2():
     #
-    from rstxml2db import main
+    from rstxml2db.cli import main
 
     result = main(['-o', 'result.xml', 'file-does-not-exist.xml'])
     assert result == ERROR_CODES[FileNotFoundError]
