+++ tests/test_version.py
@@ -3,13 +3,13 @@
 
 def test_version():
     """Check if version is available and set"""
-    from rstxml2db.version import __version__
+    from rstxml2db import __version__
     assert __version__
 
 
 def test_version_from_cli(capsys):
     """Checks if option --version creates a correct version"""
-    from rstxml2db import main
+    from rstxml2db.cli import main
     import pytest
 
     with pytest.raises(SystemExit):
